fixed decimal and float inoperability

+++ roottoplate/composter/views.py
@@ -162,17 +162,17 @@ def user_logout(request):
 
 
 def calculate_mixture_sums(cur_inputs):
-    sumN = float(sum([i['amount']*i['type'].nitrogenPercent*i['type'].moisturePercent for i in cur_inputs]))
+    sumN = sum([i['amount']*float(i['type'].nitrogenPercent*i['type'].moisturePercent) for i in cur_inputs])
     for i in cur_inputs:
-        i['carbon'] = i['type'].nitrogenPercent*i['type'].CNRatio
-    sumC = float(sum([i['amount']*i['carbon']*i['type'].moisturePercent for i in cur_inputs]))
+        i['carbon'] = float(i['type'].nitrogenPercent*i['type'].CNRatio)
+    sumC = float(sum([i['amount']*i['carbon']*float(i['type'].moisturePercent) for i in cur_inputs]))
     return sumC, sumN
 
 
 def calculate_recommended_addition(rec_input, sumC, sumN):
     cn = sumC/sumN
-    nitrogen, moisture = rec_input.get('nitrogenPercent'), rec_input.get('moisturePercent')
-    carbon = rec_input.get('nitrogenPercent')*rec_input.get('CNRatio')
+    nitrogen, moisture = float(rec_input.get('nitrogenPercent')), float(rec_input.get('moisturePercent'))
+    carbon = float(rec_input.get('nitrogenPercent')*rec_input.get('CNRatio'))
     return (cn*sumN - sumC) / (carbon*moisture - nitrogen*moisture*cn)
 
 
@@ -197,7 +197,7 @@ class InputFormView(TemplateView):
             # get the inputs without saving the form
             for input in input_formset:
                 input = input.save(commit=False)
-                cur_inputs.append({'amount': input.inputAmount, 'type': input.inputType})
+                cur_inputs.append({'amount': float(input.inputAmount), 'type': input.inputType})
             # get the total carbon and nitrogen in the mixture
             sumC, sumN = calculate_mixture_sums(cur_inputs)
             # if the ratio is too big then add more green
